Fix platform compatibility: move CompletableFuture example to JVM-only
